requirements.txt
aiohttp                       3.8.6       
aiosignal                     1.3.1       
altair                        5.1.2       
annotated-types               0.6.0       
anyio                         3.7.1       
asgiref                       3.7.2       
async-timeout                 4.0.3       
attrs                         23.1.0   
beautifulsoup4                4.12.2   
blinker                       1.7.0    
bs4                           0.0.1    
cachetools                    5.3.2    
certifi                       2023.7.22
cffi                          1.16.0   
chardet                       5.2.0
charset-normalizer            3.3.0
click                         8.1.7
colorama                      0.4.6
cryptography                  41.0.4
dataclasses-json              0.6.1
defusedxml                    0.7.1
distro                        1.8.0
Django                        4.2.6
django-allauth                0.57.0
django-allauth-2fa            0.11.1
django-axes                   6.1.1
django-cors-headers           4.3.0
django-countries              7.5.1
django-otp                    1.2.3
django-rest-authtoken         2.1.4
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.0
document                      1.0
docx                          0.2.4
et-xmlfile                    1.1.0
exceptiongroup                1.1.3
faiss-cpu                     1.7.4
fastapi                       0.104.1
filelock                      3.13.1
frozenlist                    1.4.0
fsspec                        2023.10.0
gitdb                         4.0.11
GitPython                     3.1.40
greenlet                      3.0.1
h11                           0.14.0
httpcore                      1.0.1
httpx                         0.25.1
huggingface-hub               0.17.3
idna                          3.4
importlib-metadata            6.8.0
iniconfig                     2.0.0
Jinja2                        3.1.2
joblib                        1.3.2
jsonpatch                     1.33
jsonpointer                   2.4
jsonschema                    4.19.2
jsonschema-specifications     2023.7.1
langchain                     0.0.332
langsmith                     0.0.62
lxml                          4.9.3
markdown-it-py                3.0.0
MarkupSafe                    2.1.3
marshmallow                   3.20.1
mdurl                         0.1.2
mpmath                        1.3.0
multidict                     6.0.4
mypy-extensions               1.0.0
mysqlclient                   2.2.0
networkx                      3.2.1
nltk                          3.8.1
numpy                         1.26.1
oauthlib                      3.2.2
openai                        1.2.0
openpyxl                      3.1.2
packaging                     23.2
pandas                        2.1.2
Pillow                        10.1.0
pip                           23.3.1
pluggy                        1.3.0
protobuf                      4.25.0
psycopg2-binary               2.9.9
pyarrow                       14.0.1
pycparser                     2.21
pydantic                      2.4.2
pydantic_core                 2.10.1
pydeck                        0.8.1b0
Pygments                      2.16.1
PyJWT                         2.8.0
pypdf                         3.17.1
pypng                         0.20220715.0
pytest                        7.4.3
pytest-subtests               0.11.0
python-dateutil               2.8.2
python-docx                   1.1.0
python-dotenv                 1.0.0
python-pptx                   0.6.23
python3-openid                3.2.0
pytz                          2023.3.post1
PyYAML                        6.0.1
qrcode                        7.4.2
referencing                   0.30.2
regex                         2023.10.3
reportlab                     4.0.7
requests                      2.31.0
requests-oauthlib             1.3.1
rich                          13.6.0
rpds-py                       0.12.0
safetensors                   0.4.0
scikit-learn                  1.3.2
scipy                         1.11.3
sentence-transformers         2.2.2
sentencepiece                 0.1.99
setuptools                    68.2.0
six                           1.16.0
smmap                         5.0.1
sniffio                       1.3.0
soupsieve                     2.5
SQLAlchemy                    2.0.23
sqlparse                      0.4.4
starlette                     0.27.0
sympy                         1.12
tenacity                      8.2.3
threadpoolctl                 3.2.0
tiktoken                      0.5.1
tokenizers                    0.14.1
toml                          0.10.2
tomli                         2.0.1
toolz                         0.12.0
torch                         2.1.0
torchvision                   0.16.0
tornado                       6.3.3
tqdm                          4.66.1
transformers                  4.35.0
typing_extensions             4.8.0
typing-inspect                0.9.0
tzdata                        2023.3
tzlocal                       5.2
urllib3                       2.0.6
validators                    0.22.0
watchdog                      3.0.0
wheel                         0.41.2
XlsxWriter                    3.1.9
yarl                          1.9.2
zipp                          3.17.0




